client cli is refactored
