+++ internal/client/cli/sync.go
@@ -1,4 +1,4 @@
-package main
+package cli
 
 import "github.com/spf13/cobra"
